requirements.txt
flask
flask_cors
pymongo
gunicorn
python-dotenv